requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
appnope @ file:///Users/ec2-user/ci_py311/appnope_1678317440516/work
argilla==1.5.0
arxiv==1.4.7
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==22.2.0
autoresearcher==0.0.6
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backoff==2.2.1
beautifulsoup4==4.12.0
blinker==1.6.2
boto3==1.26.76
botocore==1.29.76
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.0
certifi @ file:///Users/ec2-user/ci_py311/certifi_1678321191472/work/certifi
charset-normalizer==2.0.12
chromadb==0.3.11
click==8.1.3
clickhouse-connect==0.5.16
cloudpickle==2.2.1
comm @ file:///Users/ec2-user/ci_py311/comm_1678317779525/work
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
dacite==1.6.0
dataclasses-json==0.5.7
db-dtypes==1.1.1
debugpy @ file:///Users/ec2-user/ci_py311/debugpy_1678317801115/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deeplake==3.4.3
Deprecated==1.2.13
dill==0.3.6
discord==2.2.2
discord.py==2.2.2
dnspython==2.3.0
duckdb==0.7.1
duckduckgo-search==3.8.1
entrypoints @ file:///Users/ec2-user/ci_py311/entrypoints_1678316169365/work
et-xmlfile==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.7.3
Farama-Notifications==0.0.4
fastapi==0.95.0
feedparser==6.0.10
filelock==3.10.6
fonttools==4.39.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.9.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
gymnasium==0.28.1
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.13.3
humbug==0.3.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
ipykernel @ file:///Users/ec2-user/ci_py311/ipykernel_1678318220696/work
ipython @ file:///Users/ec2-user/ci_py311/ipython_1678317876128/work
jax-jumpy==1.0.0
jedi @ file:///Users/ec2-user/ci_py311_2/jedi_1679336327335/work
jellyfish==0.11.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client @ file:///Users/ec2-user/ci_py311/jupyter_client_1678316948346/work
jupyter_core @ file:///Users/ec2-user/ci_py311/jupyter_core_1678316236172/work
kiwisolver==1.4.4
langchain==0.0.172
langchainplus-sdk==0.0.20
libmagic==1.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline @ file:///Users/ec2-user/ci_py311/matplotlib-inline_1678317544666/work
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio @ file:///Users/ec2-user/ci_py311/nest-asyncio_1678316288195/work
networkx==3.0
nltk==3.8.1
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c3zlhpboab/croot/packaging_1678965318579/work
pandas==1.5.3
pandas-gbq==0.19.1
pandas-stubs==1.5.3.230321
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathos==0.3.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pinecone-client==2.2.1
platformdirs @ file:///Users/ec2-user/ci_py311/platformdirs_1678315988118/work
plotly==5.13.1
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit @ file:///Users/ec2-user/ci_py311/prompt-toolkit_1678317707969/work
proto-plus==1.22.2
protobuf==4.22.1
psutil @ file:///Users/ec2-user/ci_py311_2/psutil_1679337242143/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pydata-google-auth==1.7.0
pydeck==0.8.1b0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.7.0
Pympler==1.0.1
PyMuPDF==1.22.2
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-docx==0.8.11
python-dotenv==1.0.0
python-magic-bin==0.4.14
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq @ file:///Users/ec2-user/ci_py311/pyzmq_1678316754714/work
redis==3.5.3
redisearch==2.1.1
regex==2023.3.22
rejson==0.5.6
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sgmllib3k==1.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4
SQLAlchemy==1.4.47
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.26.1
streamlit==1.23.1
sympy==1.11.1
tenacity==8.2.2
termcolor==1.1.0
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado @ file:///Users/ec2-user/ci_py311/tornado_1678316833426/work
tqdm==4.65.0
traitlets @ file:///Users/ec2-user/ci_py311/traitlets_1678316097905/work
transformers==4.27.3
types-pytz==2022.7.1.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.5.7
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==10.4
wikipedia==1.4.0
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
zstandard==0.20.0
